Always open lock file readonly.
This allows locking on directories, which can't be opened in readonly.

The original mode code has been there for a long time, but according to
the flock man page, isn't needed: "A shared or exclusive lock can be
placed on a file regardless of the mode in which the file was opened."

Fixes #11.